compile with dotnet sdk rather than roslyn, use .cs files and .csproj layout in dotnet_quests
