+++ zone/dotnet_quests/.gitignore
@@ -1,3 +1,5 @@
 .DS_Store
 .vs
 obj
+.mono
+!questinterface.sln
