+++ zone/dotnet_quests/example/dotnet_quests/.gitignore
@@ -1,0 +1,5 @@
+out
+.mono
+obj
+!eq.sln
+!.vscode
