+++ zone/dotnet_quests/example/dotnet_quests/.vscode/launch.json
@@ -1,0 +1,11 @@
+{
+    "configurations": [
+        {
+            "name": ".NET Core Attach",
+            "type": "coreclr",
+            "request": "attach",
+            "processId": "${command:pickProcess}",
+            "justMyCode": false,
+        }
+    ]
+}
