+++ zone/dotnet_quests/questinterface/DotNetTypes/DotNetTypes.csproj
@@ -6,8 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <EnableDynamicLoading>true</EnableDynamicLoading>
-    <InternalsVisibleTo>RoslynBridge</InternalsVisibleTo>
-    <InternalsVisibleTo>RoslynCompiler</InternalsVisibleTo>
+    <InternalsVisibleTo>DotNetBridge</InternalsVisibleTo>
+    <InternalsVisibleTo>DotNetCompiler</InternalsVisibleTo>
     
     <!-- Use invariant globalization -->
     <InvariantGlobalization>true</InvariantGlobalization>
@@ -16,8 +16,6 @@
 
   <ItemGroup>
     <PackageReference Include="DllExport" Version="1.7.4" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.9.2" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.9.2" />
     <PackageReference Include="Microsoft.NETCore.DotNetAppHost" Version="8.0.3" />
   </ItemGroup>
