+++ zone/dotnet_quests/questinterface/questinterface.sln
@@ -1,0 +1,32 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetTypes", "DotNetTypes\DotNetTypes.csproj", "{902A540B-C23C-42B4-95E3-1FB7CB5606E0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetBridge", "DotNetBridge\DotNetBridge.csproj", "{36D371C6-5C2F-4536-93A8-362663DBA692}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{25F46D39-5CEF-4615-86C5-D9D4B483A1B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{25F46D39-5CEF-4615-86C5-D9D4B483A1B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{25F46D39-5CEF-4615-86C5-D9D4B483A1B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{25F46D39-5CEF-4615-86C5-D9D4B483A1B0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{902A540B-C23C-42B4-95E3-1FB7CB5606E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{902A540B-C23C-42B4-95E3-1FB7CB5606E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{902A540B-C23C-42B4-95E3-1FB7CB5606E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{902A540B-C23C-42B4-95E3-1FB7CB5606E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36D371C6-5C2F-4536-93A8-362663DBA692}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36D371C6-5C2F-4536-93A8-362663DBA692}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36D371C6-5C2F-4536-93A8-362663DBA692}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36D371C6-5C2F-4536-93A8-362663DBA692}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
